write JSDoc comments JC-632
