+++ apps/vanilla/src/pages/task2/index.html
@@ -14,22 +14,20 @@
 
 		<main>
 
-			<div class="container">
-				<div class="container__item dice-cap">
-					<h2>Dice-cap</h2>
-					<span class="all-results"></span>
-				</div>
-				<div class="container__item first-player-cap">
-					<h2>First player</h2>
-					<span class="first-player-results"></span>
-
-				</div>
-				<div class="container__item second-player-cap">
-					<h2>Second player</h2>
-					<span class="second-player-results"></span>
-
-				</div>
-			</div>
+			<arcticle class="container">
+				<section class="container__item dice-cap">
+					<h2 class="container__header">Dice-cap</h2>
+					<span class="container__results all-results"></span>
+				</section>
+				<section class="container__item first-player-cap">
+					<h2 class="container__header">First player</h2>
+					<span class="container__results first-player-results"></span>
+				</section>
+				<section class="container__item second-player-cap">
+					<h2 class="container__header">Second player</h2>
+					<span class="container__results second-player-results"></span>
+				</section>
+			</arcticle>
 
 			<button type="button" class="button">Throw a dice!</button>
