+++ apps/vanilla/src/pages/task2/types/subscriber.ts
@@ -1,12 +1,9 @@
-
-/**
- * Subscriber.
- */
+/** Type for subscribers which will subscribe to publishers. */
 export type Subscriber<T> = {
 
 	/**
-	 * Update method.
-	 * @param message - Message.
+	 * A method that a publisher will call when some event occurs.
+	 * @param message - A message that a publisher will send to a subscriber.
 	 */
 	update: (message: T) => void;
 };
